requirements.txt
fastapi
uvicorn
jinja2
pydantic
pywebview
pyinstaller
